remove cast

+++ examples/polybench_hls/cholesky.py
@@ -17,7 +17,7 @@ def cholesky_np(A):
         for j in range(i):
             for k in range(j):
                 A[i, j] = A[i, j] - A[i, k] * A[j, k]
-            A[i, j] = int(A[i, j] / A[j, j])
+            A[i, j] = A[i, j] / A[j, j]
         # i == j
         for k in range(i):
             A[i, i] = A[i, i] - A[i, k] * A[i, k]
